// src/app/Models/ipatient.ts
export interface IPatient {

    id: number;
    name: string;
    groupID: number;
        // phoneNumber: number;
        // Address: string;
        // gender: string;
        // birthDate: Date;
        // maritalStatus: string;
        // SSN: number;
        // SSNType: string;
        // NationalId: number;
        // bloodType: string;
        // status: string;
        // photo: string;
        // Reason: string;
    
    
    }
    